control-service: improve integration tests stability
The schedule of the already deployed data job should not be changed
since as it brings some problems. Currently the job executes on every 2
minutes and this affects some of the tests that rely on the exact executions
count.

Testing Done: integration tests
